pobranie brancha Magnum
