+++ utils.h
@@ -10,5 +10,4 @@
 int numElements(const char *fileName, int linecnt);
 int numLines(const char *filename);
 int *readLine(const char *fileName, int linecnt, int n);
-
 #endif
